ultimos cambios proyecto final

+++ src/main/java/com/luciacerpap/clinica/service/impl/OdontologoService.java
@@ -1,8 +1,13 @@
 package com.luciacerpap.clinica.service.impl;
 
+import com.luciacerpap.clinica.dto.request.OdontologoRequestDto;
+import com.luciacerpap.clinica.dto.response.OdontologoResponseDto;
 import com.luciacerpap.clinica.entity.Odontologo;
+import com.luciacerpap.clinica.exception.ResourceNotFoundException;
 import com.luciacerpap.clinica.repository.IOdontologoRepository;
 import com.luciacerpap.clinica.service.IOdontologoService;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,7 +15,8 @@ import java.util.Optional;
 
 @Service
 public class OdontologoService implements IOdontologoService {
-
+    @Autowired
+    private ModelMapper modelMapper;
     private IOdontologoRepository odontologoRepository;
 
     public OdontologoService(IOdontologoRepository odontologoRepository) {
@@ -22,8 +28,14 @@ public class OdontologoService implements IOdontologoService {
     }
 
     @Override
-    public Optional<Odontologo> buscarPorId(Integer id) {
-        return odontologoRepository.findById(id);
+    public Optional<OdontologoResponseDto> buscarPorId(Integer id) {
+        Optional<Odontologo> odontologo = odontologoRepository.findById(id);
+        if (odontologo.isPresent()) {
+            OdontologoResponseDto odontologoResponseDto = modelMapper.map(odontologo.get(), OdontologoResponseDto.class);
+            return Optional.of(odontologoResponseDto);
+        } else {
+            throw new ResourceNotFoundException("Odontologo no encontrado");
+        }
     }
 
     @Override
@@ -32,13 +44,19 @@ public class OdontologoService implements IOdontologoService {
     }
 
     @Override
-    public void modificarOdontologo(Odontologo odontologo){
-        odontologoRepository.save(odontologo);
+    public void modificarOdontologo(OdontologoRequestDto odontologo){
+        Optional<OdontologoResponseDto> odontologoAux = buscarPorId(odontologo.getId());
+        odontologoRepository.save(modelMapper.map(odontologo, Odontologo.class));
     }
 
     @Override
     public void eliminarOdontologo(Integer id){
-        odontologoRepository.deleteById(id);
+        Optional<Odontologo> odontologo = odontologoRepository.findById(id);
+        if(odontologo.isPresent()){
+            odontologoRepository.deleteById(id);
+        }else {
+            throw new ResourceNotFoundException("Odontologo no encontrado");
+        }
     }
 }
